Renamed persistence options.
