+++ src/persistence/primary/ModelManager.ts
@@ -4,8 +4,8 @@ import { SecondaryEntityManager } from '../secondary/secondary-entity-manager';
 import { IModelManager } from './IModelManager';
 import { IRedisMiddleware } from './IRedisMiddleware';
 import { MULTIPLE_RESULT_QUERY_CODE, SINGLE_RESULT_QUERY_CODE, VOID_RESULT_STRING } from './LuaConstants';
-import { AntJsDeleteOptions } from './options/AntJsDeleteOptions';
-import { AntJsUpdateOptions } from './options/AntJsUpdateOptions';
+import { AntJsDeleteOptions } from './options/antjs-delete-options';
+import { AntJsUpdateOptions } from './options/antjs-update-options';
 import { PersistencyDeleteOptions } from './options/persistency-delete-options';
 import { PersistencyUpdateOptions } from './options/persistency-update-options';
 import { PrimaryEntityManager } from './PrimaryEntityManager';
