+++ src/persistence/primary/PrimaryEntityManager.ts
@@ -5,8 +5,8 @@ import { SecondaryEntityManager } from '../secondary/secondary-entity-manager';
 import { IPrimaryEntityManager } from './IPrimaryEntityManager';
 import { IRedisMiddleware } from './IRedisMiddleware';
 import { VOID_RESULT_STRING } from './LuaConstants';
-import { AntJsSearchOptions } from './options/AntJsSearchOptions';
-import { CacheMode } from './options/CacheMode';
+import { AntJsSearchOptions } from './options/antjs-search-options';
+import { CacheMode } from './options/cache-mode';
 import { PersistencyDeleteOptions } from './options/persistency-delete-options';
 import { PersistencySearchOptions } from './options/persistency-search-options';
 import { PersistencyUpdateOptions } from './options/persistency-update-options';
