+++ src/test/primary/ModelManagerTest.ts
@@ -3,8 +3,8 @@ import { Entity } from '../../model/entity';
 import { Model } from '../../model/model';
 import { IModelManager } from '../../persistence/primary/IModelManager';
 import { IPrimaryEntityManager } from '../../persistence/primary/IPrimaryEntityManager';
-import { AntJsUpdateOptions } from '../../persistence/primary/options/AntJsUpdateOptions';
-import { CacheMode } from '../../persistence/primary/options/CacheMode';
+import { AntJsUpdateOptions } from '../../persistence/primary/options/antjs-update-options';
+import { CacheMode } from '../../persistence/primary/options/cache-mode';
 import { IPrimaryQueryManager } from '../../persistence/primary/query/IPrimaryQueryManager';
 import { ISingleResultQueryManager } from '../../persistence/primary/query/ISingleResultQueryManager';
 import { AntJsModelManagerGenerator } from '../../testapi/api/generator/AntJsModelManagerGenerator';
