+++ src/test/primary/PrimaryEntityManagerTest.ts
@@ -3,10 +3,10 @@ import { Entity } from '../../model/entity';
 import { Model } from '../../model/model';
 import { IPrimaryEntityManager } from '../../persistence/primary/IPrimaryEntityManager';
 import { ModelManager } from '../../persistence/primary/ModelManager';
-import { AntJsDeleteOptions } from '../../persistence/primary/options/AntJsDeleteOptions';
-import { AntJsSearchOptions } from '../../persistence/primary/options/AntJsSearchOptions';
-import { AntJsUpdateOptions } from '../../persistence/primary/options/AntJsUpdateOptions';
-import { CacheMode } from '../../persistence/primary/options/CacheMode';
+import { AntJsDeleteOptions } from '../../persistence/primary/options/antjs-delete-options';
+import { AntJsSearchOptions } from '../../persistence/primary/options/antjs-search-options';
+import { AntJsUpdateOptions } from '../../persistence/primary/options/antjs-update-options';
+import { CacheMode } from '../../persistence/primary/options/cache-mode';
 import { PrimaryEntityManager } from '../../persistence/primary/PrimaryEntityManager';
 import { SecondaryEntityManager } from '../../persistence/secondary/secondary-entity-manager';
 import { ITest } from '../../testapi/api/ITest';
